move logrus logging package to subfolder
