+++ logging/log.go
@@ -5,11 +5,6 @@ import (
 	"io"
 )
 
-const (
-	FIELD_SERVICE = "service"
-	FIELD_ERROR   = "error"
-)
-
 type Logger interface {
 	SimpleLog
 	FormatedLog
@@ -46,56 +41,4 @@ type CustomLog interface {
 	WithError(err error) Logger
 	WithContext(ctx context.Context) Logger
 	WriterLevel(level string) io.Writer
-}
-
-type LogMessage struct {
-	Timestamp      string `json:"timestamp"`
-	Message        string `json:"message"`
-	Severity       string `json:"severity"`
-	ReportLocation struct {
-		FilePath     string `json:"filePath,omitempty"`
-		LineNumber   int    `json:"lineNumber,omitempty"`
-		FunctionName string `json:"functionName,omitempty"`
-	} `json:"reportLocation,omitempty"`
-	Service  interface{}            `json:"service,omitempty"`
-	Metadata map[string]interface{} `json:"metadata,omitempty"`
-}
-
-type LogParam struct {
-	AppCtxEnabled bool
-	AppName       string
-	AppVersion    string
-
-	DebuggingEnabled   bool
-	PrettyPrintEnabled bool
-
-	StackSkip []string
-}
-
-type LogOption func(*LogParam)
-
-func WithAppContext(name, version string) LogOption {
-	return func(lo *LogParam) {
-		lo.AppCtxEnabled = true
-		lo.AppName = name
-		lo.AppVersion = version
-	}
-}
-
-func EnableDebugging() LogOption {
-	return func(lo *LogParam) {
-		lo.DebuggingEnabled = true
-	}
-}
-
-func EnablePrettyPrint() LogOption {
-	return func(lo *LogParam) {
-		lo.PrettyPrintEnabled = true
-	}
-}
-
-func AddStackSkip(pkg string) LogOption {
-	return func(lo *LogParam) {
-		lo.StackSkip = append(lo.StackSkip, pkg)
-	}
 }
